# docs/api-docs.yaml
openapi: 3.1.0
info:
  title: CoActivity API (core-service)
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /api/users/me:
    get:
      tags:
      - Users
      summary: Профиль текущего пользователя
      operationId: getUserProfile
      responses:
        "200":
          description: Профиль пользователя.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "401":
          description: Требуется аутентификация.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "500":
          description: Неожиданная ошибка.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
      security:
      - sessionCookie: []
    put:
      tags:
      - Users
      summary: Обновить профиль текущего пользователя
      description: Обновляет поля профиля. Требуются session cookie и CSRF.
      operationId: updateUserProfile
      parameters:
      - name: X-XSRF-TOKEN
        in: header
        description: CSRF токен (дублирует cookie XSRF-TOKEN).
        required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdateRequest"
        required: true
      responses:
        "200":
          description: Обновлённый профиль.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "400":
          description: Ошибка валидации.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "401":
          description: Требуется аутентификация.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "500":
          description: Неожиданная ошибка.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
      security:
      - sessionCookie: []
    delete:
      tags:
      - Users
      summary: Удалить аккаунт (если нет комнат-владений)
      description: "Удаляет аккаунт текущего пользователя, только если у него нет\
        \ комнат, где он OWNER. Требуются session cookie и CSRF. Выполняет logout\
        \ и инвалидирует сессии."
      operationId: deleteAccount
      parameters:
      - name: X-XSRF-TOKEN
        in: header
        description: CSRF токен (дублирует cookie XSRF-TOKEN).
        required: true
      responses:
        "204":
          description: Аккаунт удалён.
        "401":
          description: Требуется аутентификация.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "403":
          description: "Удаление запрещено (например, есть owned rooms)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "409":
          description: Конфликт состояния.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "500":
          description: Неожиданная ошибка.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
      security:
      - sessionCookie: []
...
  /api/auth/register:
    post:
      tags:
      - Auth
      summary: Регистрация пользователя
      description: "Создаёт пользователя и инициирует подтверждение регистрации. Д\
        ля POST требуется CSRF: сначала вызови GET /api/auth/csrf и передай заголо\
        вок X-XSRF-TOKEN."
      operationId: register
      parameters:
      - name: X-XSRF-TOKEN
        in: header
        description: CSRF токен (дублирует cookie XSRF-TOKEN).
        required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistrationRequest"
        required: true
      responses:
        "201":
          description: Пользователь создан.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResponse"
        "400":
          description: Ошибка валидации входных данных.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "409":
          description: "Конфликт (например, email уже зарегистрирован)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "500":
          description: Неожиданная ошибка.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
  /api/auth/register/verify:
    post:
      tags:
      - Auth
      summary: Подтвердить регистрацию
      description: "Подтверждает регистрацию по коду. Для POST требуется CSRF: сна\
        чала вызови GET /api/auth/csrf и передай заголовок X-XSRF-TOKEN."
      operationId: verifyRegistration
      parameters:
      - name: X-XSRF-TOKEN
        in: header
        description: CSRF токен (дублирует cookie XSRF-TOKEN).
        required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterVerificationRequest"
        required: true
      responses:
        "204":
          description: Регистрация подтверждена.
        "400":
          description: Ошибка валидации.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "404":
          description: Код/пользователь не найден.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "409":
          description: "Конфликт (например, уже подтверждено)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "500":
          description: Неожиданная ошибка.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
  /api/auth/register/resend:
    post:
      tags:
      - Auth
      summary: Переотправить код подтверждения регистрации
      description: "Переотправляет код подтверждения. Для POST требуется CSRF: сна\
        чала вызови GET /api/auth/csrf и передай заголовок X-XSRF-TOKEN."
      operationId: resendRegistrationCode
      parameters:
      - name: X-XSRF-TOKEN
        in: header
        description: CSRF токен (дублирует cookie XSRF-TOKEN).
        required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendRegistrationVerificationRequest"
        required: true
      responses:
        "204":
          description: Код переотправлен.
        "400":
          description: Ошибка валидации.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "404":
          description: Пользователь/регистрация не найдены.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "500":
          description: Неожиданная ошибка.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
  /api/auth/password/reset/verify:
    post:
      tags:
      - Auth
      summary: "Password reset: проверить код"
      description: "Проверяет корректность кода сброса пароля. Для POST требуется\
        \ CSRF: сначала вызови GET /api/auth/csrf и передай заголовок X-XSRF-TOKEN."
      operationId: verifyPasswordResetCode
      parameters:
      - name: X-XSRF-TOKEN
        in: header
        description: CSRF токен (дублирует cookie XSRF-TOKEN).
        required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetVerifyRequest"
        required: true
      responses:
        "204":
          description: Код корректен.
        "400":
          description: Ошибка валидации.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "404":
          description: Код/пользователь не найдены.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "409":
          description: Конфликт/истёкший код.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "500":
          description: Неожиданная ошибка.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
  /api/auth/password/reset/request:
    post:
      tags:
      - Auth
      summary: "Password reset: запросить код"
      description: "Инициирует сброс пароля и отправляет код/ссылку пользователю.\
        \ Для POST требуется CSRF: сначала вызови GET /api/auth/csrf и передай заг\
        оловок X-XSRF-TOKEN."
      operationId: requestPasswordReset
      parameters:
      - name: X-XSRF-TOKEN
        in: header
        description: CSRF токен (дублирует cookie XSRF-TOKEN).
        required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
        required: true
      responses:
        "204":
          description: Запрос принят.
        "400":
          description: Ошибка валидации.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "404":
          description: Пользователь не найден.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "500":
          description: Неожиданная ошибка.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
  /api/auth/password/reset/confirm:
    post:
      tags:
      - Auth
      summary: "Password reset: подтвердить новый пароль"
      description: "Устанавливает новый пароль по подтверждённому коду. Для POST т\
        ребуется CSRF: сначала вызови GET /api/auth/csrf и передай заголовок X-XSRF-TOKEN."
      operationId: confirmPasswordReset
      parameters:
      - name: X-XSRF-TOKEN
        in: header
        description: CSRF токен (дублирует cookie XSRF-TOKEN).
        required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirmRequest"
        required: true
      responses:
        "204":
          description: Пароль обновлён.
        "400":
          description: Ошибка валидации.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "404":
          description: Код/пользователь не найдены.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "409":
          description: Конфликт/истёкший код.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
        "500":
          description: Неожиданная ошибка.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblemDetail"
...
components:
  schemas:
    UserProfileUpdateRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 20
          minLength: 2
        dateOfBirth:
          type: string
          format: date-time
        city:
          type: string
          maxLength: 100
          minLength: 0
        country:
          type: string
          maxLength: 100
          minLength: 0
        description:
          type: string
          maxLength: 500
          minLength: 0
        avatarId:
          type: integer
          format: int32
    UserProfileResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        username:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        city:
          type: string
        country:
          type: string
        description:
          type: string
        avatarId:
          type: integer
          format: int32
        avatarUrl:
          type: string
        notifications:
          type: array
          items:
            type: string
            enum:
            - membershipAccepted
            - membershipRejected
            - activityClosed
            - newJoinRequest
            - importantRoomUpdates
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      description: HTTP session cookie. Issued on login; sent automatically by the
        browser.
      name: COACTIVITY_SESSION
      in: cookie
    csrfToken:
      type: apiKey
      description: CSRF header value for state-changing requests. Token is exposed
        via cookie XSRF-TOKEN and must be echoed in header X-XSRF-TOKEN.
      name: X-XSRF-TOKEN
      in: header